requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
mcp==1.8.0
uvicorn>=0.32.1
mcp[http]
pan-aisecurity==0.4.0.5.post2
httpx==0.27.2
tenacity